chrismaree/further gas savings (#4285)

+++ packages/core/contracts/optimistic-asserter/implementation/examples/DataAsserter.sol
@@ -13,7 +13,7 @@ contract DataAsserter {
     using SafeERC20 for IERC20;
     IERC20 public immutable defaultCurrency;
     OptimisticAsserterInterface public immutable oa;
-    uint256 public constant assertionLiveness = 7200;
+    uint64 public constant assertionLiveness = 7200;
     bytes32 public immutable defaultIdentifier;
 
     struct DataAssertion {
