+++ packages/core/contracts/optimistic-asserter/implementation/test/OptimisticAsserterTest.sol
@@ -9,11 +9,11 @@ contract OptimisticAsserterTest is OptimisticAsserter, Testable {
     constructor(
         FinderInterface _finder,
         IERC20 _defaultCurrency,
-        uint256 _defaultLiveness,
+        uint64 _defaultLiveness,
         address _timerAddress
     ) OptimisticAsserter(_finder, _defaultCurrency, _defaultLiveness) Testable(_timerAddress) {}
 
     function getCurrentTime() public view override(OptimisticAsserter, Testable) returns (uint256) {
-        return Testable.getCurrentTime();
+        return uint256(Testable.getCurrentTime());
     }
 }
