+++ packages/core/test/foundry/optimistic-asserter/Common.sol
@@ -27,7 +27,7 @@ contract Common is Test {
     bytes falseClaimAssertion = bytes("q:'The sky is red'");
     uint256 burnedBondPercentage;
     uint256 defaultBond;
-    uint256 defaultLiveness;
+    uint64 defaultLiveness;
     bytes32 defaultIdentifier;
 
     // Mock addresses, used to prank calls.
@@ -46,7 +46,7 @@ contract Common is Test {
         address caller,
         IERC20 currency,
         uint256 bond,
-        uint256 expirationTime
+        uint64 expirationTime
     );
     event AssertionDisputed(bytes32 indexed assertionId, address indexed disputer);
